+++ src/main/java/com/feiyu/discord/sg/tavern/repositories/EventRepository.java
@@ -3,10 +3,13 @@ package com.feiyu.discord.sg.tavern.repositories;
 import com.feiyu.discord.sg.tavern.entities.EventEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface EventRepository extends JpaRepository<EventEntity, String> {
     
     Optional<EventEntity> findTopByPostId(String postId);
+    
+    List<EventEntity> findAllByPostStatus(String postStatus);
     
 }